requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
distlib==0.3.6
dm-tree==0.1.8
Farama-Notifications==0.0.4
filelock==3.12.0
flatbuffers==23.5.26
frozenlist==1.3.3
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.3
grpcio-tools==1.54.2
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h5py==3.8.0
idna==3.4
imageio==2.30.0
jax==0.4.11
jax-jumpy==1.0.0
jsonschema==4.17.3
keras==2.12.0
lazy_loader==0.2
libclang==16.0.0
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.1.0
mpi4py==3.1.4
msgpack==1.0.5
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.1
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
ray==2.4.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboardX==2.6
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tifffile==2023.4.12
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
Werkzeug==2.3.4
wrapt==1.14.1
